pango: manually implement LayoutDeserializeError
Because gir picks the wrong quark function
We also don't have a way to disable generating the error domain only for an enum.
