+++ pango/Gir.toml
@@ -34,7 +34,6 @@ generate = [
     "Pango.Gravity",
     "Pango.GravityHint",
     "Pango.Item",
-    "Pango.LayoutDeserializeError",
     "Pango.LayoutDeserializeFlags",
     "Pango.LayoutIter",
     "Pango.LayoutRun",
@@ -73,6 +72,7 @@ manual = [
     "Pango.GlyphGeometry",
     "Pango.GlyphInfo",
     "Pango.Language",
+    "Pango.LayoutDeserializeError",
     "Pango.Rectangle",
 ]
