+++ pango/src/auto/mod.rs
@@ -86,9 +86,6 @@ pub use self::enums::EllipsizeMode;
 pub use self::enums::FontScale;
 pub use self::enums::Gravity;
 pub use self::enums::GravityHint;
-#[cfg(any(feature = "v1_50", feature = "dox"))]
-#[cfg_attr(feature = "dox", doc(cfg(feature = "v1_50")))]
-pub use self::enums::LayoutDeserializeError;
 #[cfg(any(feature = "v1_46", feature = "dox"))]
 #[cfg_attr(feature = "dox", doc(cfg(feature = "v1_46")))]
 pub use self::enums::Overline;
